RDLN: don't spit NL>
When being called in VE, the NL> spit messed up the first line
of contents. Instead, we now spit the NL> in RDLN<.

Also, move RDLN and C< initialization, which was repeated in
both BOOT and QUIT, in (main). This deduplicates the words
and allow "IN"-type of words to be declared later.