Add "constant" word type
This give us two benefits: it makes access to HERE and
CURRENT, which are very often used, much faster, and it allows
the use of CONSTANT in cross-compiled code.

We can't hope to ever have a working DOES> in cross-compiled
code, but having CONSTANT removes a lot of friction by itself.