Make QUIT and ABORT native words
To make it work, I had to add (main) to the stable ABI.
Those new native words replace (resSP) and (resRS) which were
only used by QUIT and ABORT.
The goal of this change is to allow QUITing from native code.
One example I have in mind is a tight loop check for a BREAK
key press which triggers QUIT. I would need this to break free
of stuck RS-232 wait loops without having to reset the whole
system.