xcomp: implement XFIND in Forth instead of re-using _xfind
I'm on a collision course with big-endian and prev fields,
which are hardcoded in _find implementations. To avert this
I'm first re-implementing XFIND so that it performs the whole
logic and stops using the XCON/XCOFF dance.

This include all find-related words. I had a bit of a problem
with X' which was used to give ":" and "(" its names in late
xcomp because it was the only XFIND-related word that didn't
apply offset on its result. I reorganized the ":" and "("
words so that XFIND wasn't used there anymore.